requirements.txt
requests
boto3
feedparser
newspaper3k
datefinder
